Agent source cleanup
