+++ packages/agent/src/hl7.test.ts
@@ -3,7 +3,7 @@ import { Agent, Bot, Endpoint, Resource } from '@medplum/fhirtypes';
 import { Hl7Client, Hl7Server } from '@medplum/hl7';
 import { MockClient } from '@medplum/mock';
 import { Client, Server } from 'mock-socket';
-import { App } from './main';
+import { App } from './app';
 
 jest.mock('node-windows');
 
@@ -25,41 +25,6 @@ describe('Agent', () => {
       resourceType: 'Endpoint',
       address: 'mllp://0.0.0.0:57000',
     });
-  });
-
-  test('Runs successfully', async () => {
-    const mockServer = new Server('wss://example.com/ws/agent');
-
-    mockServer.on('connection', (socket) => {
-      socket.on('message', (data) => {
-        const command = JSON.parse((data as Buffer).toString('utf8'));
-        if (command.type === 'connect') {
-          socket.send(
-            Buffer.from(
-              JSON.stringify({
-                type: 'connected',
-              })
-            )
-          );
-        }
-      });
-    });
-
-    const agent = await medplum.createResource<Agent>({
-      resourceType: 'Agent',
-      channel: [
-        {
-          endpoint: createReference(endpoint),
-          targetReference: createReference(bot),
-        },
-      ],
-    });
-
-    const app = new App(medplum, agent.id as string);
-    await app.start();
-    app.stop();
-    app.stop();
-    mockServer.stop();
   });
 
   test('Send and receive', async () => {
